% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2khabmon.R
\docType{package}
\name{n2khabmon-package}
\alias{n2khabmon}
\alias{n2khabmon-package}
\title{n2khabmon: Prepare and Manage N2KHAB Monitoring Schemes}
\description{
n2khabmon is an R package with utilities to prepare and manage Flemish monitoring schemes regarding Natura 2000 habitats and regionally important biotopes (RIBs).
}
\details{
See the \href{00Index.html}{Index}
for documentation on functions and datasets delivered with the package.

For contributing, see the README on
\href{https://github.com/inbo/n2khabmon/blob/master/README.md}{GitHub}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://inbo.github.io/n2khabmon}
  \item \url{https://github.com/inbo/n2khabmon}
  \item Report bugs at \url{https://github.com/inbo/n2khabmon/issues}
}

}
\author{
\strong{Maintainer}: Floris Vanderhaeghe \email{floris.vanderhaeghe@inbo.be} (\href{https://orcid.org/0000-0002-6378-6229}{ORCID})

Other contributors:
\itemize{
  \item Toon Westra \email{toon.westra@inbo.be} (\href{https://orcid.org/0000-0003-2478-9459}{ORCID}) [contributor]
  \item Research Institute for Nature and Forest (INBO) \email{info@inbo.be} [copyright holder]
}

}
\keyword{internal}
